pyproject.toml
[project]
name = "flame"
version = "0.1.0"
description = "AI-powered CLI coding assistant with Hack Club AI"
requires-python = ">=3.10"
dependencies = [
    "openrouter>=0.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "prompt-toolkit>=3.0.0",
]

[project.scripts]
flame = "flame.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"]}}

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

